+++ character/main_character.gd
@@ -24,5 +24,4 @@ func _physics_process(delta):
 		move_and_slide()
 
 	if Input.is_action_pressed('attack'):
-		is_attacking = true
-		_attack()
+		_start_attack(get_global_mouse_position())